(#3550) Add DotNetVSTest alias (synonym to DotNetCoreVSTest)
